refactor: github oauth client 수정
