--- backend/src/main/java/develup/infra/auth/github/GithubOAuthProperties.java
@@ -1,20 +1,0 @@
-package develup.infra.auth.github;
-
-import jakarta.validation.constraints.NotBlank;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-
-@ConfigurationProperties("auth.github")
-public record GithubOAuthProperties(
-        @NotBlank
-        String clientId,
-
-        @NotBlank
-        String clientSecret,
-
-        @NotBlank
-        String redirectUri,
-
-        @NotBlank
-        String clientUri
-) {
-}
