+++ backend/src/main/java/develup/infra/auth/oauth/github/GithubOAuthProperties.java
@@ -1,0 +1,13 @@
+package develup.infra.auth.oauth.github;
+
+import jakarta.validation.constraints.NotBlank;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+
+@ConfigurationProperties("auth.github")
+public record GithubOAuthProperties(
+        @NotBlank String clientId,
+        @NotBlank String clientSecret,
+        @NotBlank String redirectUri,
+        @NotBlank String clientUri
+) {
+}
