+++ backend/src/main/java/develup/infra/auth/oauth/github/dto/GithubAccessTokenResponse.java
@@ -1,4 +1,4 @@
-package develup.infra.auth.github.dto;
+package develup.infra.auth.oauth.github.dto;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies.SnakeCaseStrategy;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
